Add certificates and modify serialization-related code
