+++ 2-advanced/dubbo-samples-spring-security/dubbo-samples-spring-security-common/pom.xml
@@ -33,7 +33,13 @@
         <!-- dubbo -->
         <dependency>
             <groupId>org.apache.dubbo</groupId>
-            <artifactId>dubbo-spring-security</artifactId>
+            <artifactId>dubbo</artifactId>
+        </dependency>
+
+        <!-- springboot -->
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter</artifactId>
         </dependency>
 
     </dependencies>
